requirements.txt
Django==1.10.1
psycopg2
django-npm
django-compressor
py-pretty